requirements.txt
python-telegram-bot==21.4
gspread==6.1.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
google-api-python-client==2.135.0
aiosqlite==0.20.0
